refactored duplicate test helper
