+++ pkg/apis/sources/v1alpha1/apiserver_conversion_test.go
@@ -18,7 +18,7 @@ package v1alpha1
 
 import (
 	"context"
-	"errors"
+	"knative.dev/eventing/pkg/apis/sources"
 	"reflect"
 	"testing"
 
@@ -35,20 +35,8 @@ import (
 	duckv1 "knative.dev/pkg/apis/duck/v1"
 )
 
-// TODO: Replace dummy some other Eventing object once they
-// implement apis.Convertible
-type dummy struct{}
-
-func (*dummy) ConvertTo(ctx context.Context, obj apis.Convertible) error {
-	return errors.New("Won't go")
-}
-
-func (*dummy) ConvertFrom(ctx context.Context, obj apis.Convertible) error {
-	return errors.New("Won't go")
-}
-
 func TestApiServerSourceConversionBadType(t *testing.T) {
-	good, bad := &ApiServerSource{}, &dummy{}
+	good, bad := &ApiServerSource{}, &sources.DummyObject{}
 
 	if err := good.ConvertTo(context.Background(), bad); err == nil {
 		t.Errorf("ConvertTo() = %#v, wanted error", bad)
