+++ pkg/apis/sources/v1alpha2/apiserver_conversion_test.go
@@ -18,6 +18,7 @@ package v1alpha2
 
 import (
 	"context"
+	"knative.dev/eventing/pkg/apis/sources"
 	"reflect"
 	"testing"
 
@@ -31,7 +32,7 @@ import (
 )
 
 func TestApiServerSourceConversionBadType(t *testing.T) {
-	good, bad := &ApiServerSource{}, &dummy{}
+	good, bad := &ApiServerSource{}, &sources.DummyObject{}
 
 	if err := good.ConvertTo(context.Background(), bad); err == nil {
 		t.Errorf("ConvertTo() = %#v, wanted error", bad)
