+++ pkg/apis/sources/v1alpha2/container_conversion_test.go
@@ -18,6 +18,7 @@ package v1alpha2
 
 import (
 	"context"
+	"knative.dev/eventing/pkg/apis/sources"
 	"reflect"
 	"testing"
 
@@ -31,7 +32,7 @@ import (
 )
 
 func TestContainerSourceConversionBadType(t *testing.T) {
-	good, bad := &ContainerSource{}, &dummy{}
+	good, bad := &ContainerSource{}, &sources.DummyObject{}
 
 	if err := good.ConvertTo(context.Background(), bad); err == nil {
 		t.Errorf("ConvertTo() = %#v, wanted error", bad)
