+++ pkg/apis/sources/v1alpha2/ping_conversion_test.go
@@ -18,6 +18,7 @@ package v1alpha2
 
 import (
 	"context"
+	"knative.dev/eventing/pkg/apis/sources"
 	"reflect"
 	"testing"
 
@@ -29,7 +30,7 @@ import (
 )
 
 func TestPingSourceConversionBadType(t *testing.T) {
-	good, bad := &PingSource{}, &PingSource{}
+	good, bad := &PingSource{}, &sources.DummyObject{}
 
 	if err := good.ConvertTo(context.Background(), bad); err == nil {
 		t.Errorf("ConvertTo() = %#v, wanted error", bad)
