+++ pkg/apis/sources/v1beta1/apiserver_conversion_test.go
@@ -18,7 +18,7 @@ package v1beta1
 
 import (
 	"context"
-	"errors"
+	"knative.dev/eventing/pkg/apis/sources"
 	"reflect"
 	"testing"
 
@@ -45,19 +45,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// implement apis.Convertible
-type dummy struct{}
-
-func (*dummy) ConvertTo(ctx context.Context, obj apis.Convertible) error {
-	return errors.New("Won't go")
-}
-
-func (*dummy) ConvertFrom(ctx context.Context, obj apis.Convertible) error {
-	return errors.New("Won't go")
-}
-
 func TestApiServerSourceConversionBadType(t *testing.T) {
-	good, bad := &ApiServerSource{}, &dummy{}
+	good, bad := &ApiServerSource{}, &sources.DummyObject{}
 
 	if err := good.ConvertTo(context.Background(), bad); err == nil {
 		t.Errorf("ConvertTo() = %#v, wanted error", bad)
