+++ pkg/apis/sources/v1beta1/sinkbinding_conversion_test.go
@@ -18,11 +18,12 @@ package v1beta1
 
 import (
 	"context"
+	"knative.dev/eventing/pkg/apis/sources"
 	"testing"
 )
 
 func TestSinkBindingConversionBadType(t *testing.T) {
-	good, bad := &SinkBinding{}, &SinkBinding{}
+	good, bad := &SinkBinding{}, &sources.DummyObject{}
 
 	if err := good.ConvertTo(context.Background(), bad); err == nil {
 		t.Errorf("ConvertTo() = %#v, wanted error", bad)
